chore(KCP IpRange): preparations for implementation v2 and optional cidr
