+++ pkg/kcp/provider/aws/iprange/v1/updateSuccessStatus.go
@@ -1,4 +1,4 @@
-package iprange
+package v1
 
 import (
 	"context"
