+++ pkg/kcp/provider/aws/iprange/v1/util.go
@@ -1,4 +1,4 @@
-package iprange
+package v1
 
 const (
 	tagKey = "cloud-manager.kyma-project.io/iprange"
